Convert to use BigDecimal

+++ CryptoData/app/src/main/java/dev/shorthouse/cryptodata/data/source/remote/model/CoinChartApiModel.kt
@@ -1,5 +1,7 @@
 package dev.shorthouse.cryptodata.data.source.remote.model
 
+import java.math.BigDecimal
+
 data class CoinChartApiModel(
-    val prices: List<List<Double>>
+    val prices: List<List<BigDecimal>>
 )
